Disable support for systems needing 64-bit indices
The 64-bit support isn't implemented everywhere and probably would exceed
memory limits even if it was. BigUINT has been set to 32 bit and a check
is added to cudaMain to confirm that this is large enough for the system.
If it's not, the program will exit with a failure status.

+++ include/Datatypes.h
@@ -43,7 +43,8 @@ typedef float4 Real4;
 typedef float2 Complex;
 
 #endif
-typedef uint64_t BigUINT;
+// this will probably not work properly if BigUINT is set to uint64_t
+typedef uint32_t BigUINT;
 typedef uint32_t UINT;
 
 #define NUM_THREADS 128
